+++ NLBE-Bot/Blitzstars/All.cs
@@ -1,0 +1,73 @@
+namespace NLBE_Bot.Blitzstars;
+
+public class All
+{
+	public int spotted
+	{
+		get; set;
+	}
+	public int hits
+	{
+		get; set;
+	}
+	public int frags
+	{
+		get; set;
+	}
+	public int max_xp
+	{
+		get; set;
+	}
+	public int wins
+	{
+		get; set;
+	}
+	public int losses
+	{
+		get; set;
+	}
+	public int capture_points
+	{
+		get; set;
+	}
+	public int battles
+	{
+		get; set;
+	}
+	public int damage_dealt
+	{
+		get; set;
+	}
+	public int damage_received
+	{
+		get; set;
+	}
+	public int max_frags
+	{
+		get; set;
+	}
+	public int shots
+	{
+		get; set;
+	}
+	public int frags8p
+	{
+		get; set;
+	}
+	public int xp
+	{
+		get; set;
+	}
+	public int win_and_survived
+	{
+		get; set;
+	}
+	public int survived_battles
+	{
+		get; set;
+	}
+	public int dropped_capture_points
+	{
+		get; set;
+	}
+}
