+++ NLBE-Bot/Blitzstars/Vehicle.cs
@@ -1,0 +1,49 @@
+namespace NLBE_Bot.Blitzstars;
+
+public class Vehicle
+{
+	public All all
+	{
+		get; set;
+	}
+	public int last_battle_time
+	{
+		get; set;
+	}
+	public int account_id
+	{
+		get; set;
+	}
+	public int max_xp
+	{
+		get; set;
+	}
+	public int in_garage_updated
+	{
+		get; set;
+	}
+	public int max_frags
+	{
+		get; set;
+	}
+	public string frags
+	{
+		get; set;
+	}
+	public int mark_of_mastery
+	{
+		get; set;
+	}
+	public int battle_life_time
+	{
+		get; set;
+	}
+	public string in_garage
+	{
+		get; set;
+	}
+	public int tank_id
+	{
+		get; set;
+	}
+}
